+++ src/TwigComponent/src/CVA.php
@@ -24,6 +24,8 @@ namespace Symfony\UX\TwigComponent;
  * @author Mathéo Daninos <matheo.daninos@gmail.com>
  *
  * @experimental
+ *
+ * @deprecated since TwigComponent 2.20, use CVA from the "twig/html-extra:^3.12.0" package instead.
  */
 final class CVA
 {